refactor: move sortable fields in People
